AST: Move some code from ProtocolConformance.cpp to Effects.cpp / TypeCheckEffects.cpp
